pyproject.toml
[project]
name = "flamerelay"
version = "0.1.0"
description = "Silly Django app for tracking lighters"
readme = "README.md"
license = { text = "MIT" }
authors = [ { name = "jvacek", email = "contact@jvacek.eu" } ]
requires-python = "==3.14.*"
classifiers = [
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.14",
]
dependencies = [
  "argon2-cffi==25.1.0",
  "captcha>=0.7.1",
  "celery==5.6.3",
  "crispy-bootstrap5==2026.3",
  "django==6.0.4",
  "django-allauth[mfa]==65.16.1",
  "django-anymail[mailtrap]==15.0",
  "django-case-insensitive-field>=1.0.7",
  "django-celery-beat==2.9.0",
  "django-cors-headers==4.9.0",
  "django-crispy-forms==2.6",
  "django-csp>=4.0",
  "django-environ==0.13.0",
  "django-model-utils==5.0.0",
  "django-redis==6.0.0",
  "django-resized>=1.0.3",
  "django-simple-captcha>=0.6.3",
  "django-storages==1.14.6",
  "django-webpack-loader==3.2.3",
  "djangorestframework==3.17.1",
  "djangorestframework-gis>=1.0",
  "drf-spectacular==0.29.0",
  "flower==2.0.1",
  "geopy>=2.4.1",
  "gunicorn==25.3.0",
  "hiredis==3.3.1",
  "pillow==12.2.0",
  "psycopg[c]==3.3.4",
  "pyjwt>=2.12.1",
  "python-slugify==8.0.4",
  "redis==7.4.0",
  "sentry-sdk==2.58.0",
  "webauthn>=2.0",
  "whitenoise==6.12.0",
]

[dependency-groups]
dev = [
  "coverage==7.13.5",
  "django-coverage-plugin==3.2.2",
  "django-debug-toolbar==6.3.0",
  "django-extensions==4.1",
  "django-fastdev>=1.13.0",
  "django-migrations-ruff-formatter>=0.1.0",
  "django-stubs>=6.0.3",
  "django-template-lsp>=1.2.2",
  "djangorestframework-stubs==3.16.9",
  "djlint==1.36.4",
  "factory-boy==3.3.3",
  "ipdb==0.13.13",
  # "mypy==1.20.0",
  # "pre-commit==4.5.1",
  "psycopg[c]==3.3.4",
  "pyrefly>=0.63.1",
  "pytest==9.0.3",
  "pytest-django==4.12.0",
  "pytest-sugar==1.1.1",
  "ruff==0.15.12",
  "sphinx==9.1.0",
  "sphinx-autobuild==2025.8.25",
  "watchfiles==1.1.1",
  "werkzeug[watchdog]==3.1.8",
]

[tool.ruff]
line-length = 120
# Exclude a variety of commonly ignored directories.
extend-exclude = [ "staticfiles/*" ]
lint.select = [
  "A",
  "ASYNC",
  "B",
  "BLE",
  "C4",
  "C90",
  "COM",
  "DJ",
  "DTZ",
  "E",
  "EM",
  # "ERA",
  "EXE",
  "F",
  "FA",
  "FBT",
  "FLY",
  "G",
  "I",
  "ICN",
  "INP",
  "INT",
  "ISC",
  "N",
  "PD",
  "PERF",
  "PGH",
  "PIE",
  "PL",
  "PT",
  "PTH",
  "PYI",
  "Q",
  "RET",
  "RSE",
  "RUF",
  "S",
  "SIM",
  "SLF",
  "SLOT",
  "T10",
  "T20",
  "TC",
  "TID",
  "TRY",
  "UP",
  "W",
  "YTT",
]
lint.ignore = [
  # of types for comparison.
  # Deactivated because it can make the code slow:
  # https://github.com/astral-sh/ruff/issues/7871
  "COM812",
  "RUF012", # Mutable class attributes should be annotated with `typing.ClassVar`
  "S101",   # Use of assert detected https://docs.astral.sh/ruff/rules/assert/
  "SIM102", # sometimes it's better to nest
]

# lint.isort.force-single-line = true
[tool.pyproject-fmt]
keep_full_version = true

[tool.pyrefly]
python-version = "3.14.0"
ignore-missing-imports = [ "*" ]

[tool.django-stubs]
django_settings_module = "config.settings.test"

[tool.pytest]
# ==== pytest ====
ini_options.minversion = "6.0"
ini_options.addopts = "--ds=config.settings.test --reuse-db --import-mode=importlib"
ini_options.python_files = [ "tests.py", "test_*.py" ]

[tool.coverage]
# ==== Coverage ====
run.include = [ "flamerelay/**" ]
run.omit = [ "*/migrations/*", "*/tests/*" ]
run.plugins = [ "django_coverage_plugin" ]

# ==== djLint ====
[tool.djlint]
blank_line_after_tag = "load,extends"
close_void_tags = true
format_css = true
format_js = true
# TODO: remove T002 when fixed https://github.com/djlint/djLint/issues/687
ignore = "H006,H021,H030,H031,T002,D018"
include = "H017,H035"
indent = 2
max_line_length = 119
profile = "django"
css.indent_size = 2
js.indent_size = 2
